[BE] fix : 상담 예약 정보 조회 분리
- 유저와 상담사가 각각 확인할 수 있도록 메서드 분리

+++ BE/src/main/java/com/ssafy/inmind/reservation/dto/ReserveResponseDto.java
@@ -12,7 +12,7 @@ import java.time.LocalTime;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class ReserveResponseDto {
-    private int reserveInfoIdx;
+    private Long reserveInfoIdx;
     private String coName;
     private LocalDate reserveInfoDate;
     private LocalTime reserveInfoStartTime;
@@ -20,7 +20,7 @@ public class ReserveResponseDto {
 
     public static ReserveResponseDto fromEntity(Reservation reservation) {
         return ReserveResponseDto.builder()
-                .reserveInfoIdx((int) reservation.getId())
+                .reserveInfoIdx(reservation.getId())
                 .coName(reservation.getCounselor().getName())
                 .reserveInfoDate(reservation.getLocalDate())
                 .reserveInfoStartTime(reservation.getStartTime())
